+++ client/src/components/search/index.ts
@@ -1,1 +1,2 @@
 export { default as SearchPanel } from "./SearchPanel";
+export { default as SearchSyntaxHelp } from "./SearchSyntaxHelp";